+++ src/main/java/nextstep/member/AuthConfig.java
@@ -1,7 +1,7 @@
 package nextstep.member;
 
 import nextstep.member.application.JwtTokenProvider;
-import nextstep.member.ui.AuthenticationPrincipalArgumentResolver;
+import nextstep.member.presentation.AuthenticationPrincipalArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
