+++ src/main/java/nextstep/member/presentation/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package nextstep.member.ui;
+package nextstep.member.presentation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
