+++ src/main/java/nextstep/member/presentation/MemberController.java
@@ -1,4 +1,4 @@
-package nextstep.member.ui;
+package nextstep.member.presentation;
 
 import nextstep.member.application.MemberService;
 import nextstep.member.application.dto.MemberRequest;
